Move ship design cache to a separate file
