+++ default/python/AI/ship_design/__init__.py
@@ -1,1 +1,3 @@
 from ship_design.additional_specification import AdditionalSpecifications
+from ship_design.design_tools import recursive_dict_diff, build_reference_name
+from ship_design.ship_design_cache import ShipDesignCache
